Filters saved and loading screen created
Filters are now saved when the Start Program button is pressed from the selector menu. This then takes them to the new Loading Screen, which visually states all the filters being applied. Altair installed and ready for integration(in theory). Start program now spits an error if no primary archetype is selected